remote-wallet: Move `Locator` to its own module
