+++ remote-wallet/src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(clippy::integer_arithmetic)]
 pub mod ledger;
 pub mod ledger_error;
+pub mod locator;
 pub mod remote_keypair;
 pub mod remote_wallet;
